change: fix: moved bridge alerts to separate hooks
