+++ packages/boba/gateway/src/containers/Bridging/BridgeInput/index.tsx
@@ -1,20 +1,13 @@
-import { openModal } from 'actions/uiAction'
-import React, { FC, useEffect } from 'react'
+import React, { FC } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
-import {
-  selectAccountEnabled,
-  selectBridgeType,
-  selectLayer,
-  selectTokenToBridge,
-} from 'selectors'
+import { openModal } from 'actions/uiAction'
+import { selectAccountEnabled, selectTokenToBridge } from 'selectors'
 
-import { clearBridgeAlert, setBridgeAlert } from 'actions/bridgeAction'
 import useBridgeSetup from 'hooks/useBridgeSetup'
 import { getCoinImage } from 'util/coinImage'
-import { LAYER } from 'util/constant'
 
 import useAmountToReceive from 'hooks/useAmountToReceive'
-import { BRIDGE_TYPE } from '../BridgeTypeSelector'
+import useBridgeAlerts from 'hooks/useBridgeAlerts'
 import { Label } from '../styles'
 import BridgeToAddress from './BridgeToAddress'
 import Fee from './Fee'
@@ -39,37 +32,9 @@ const BridgeInput: FC<Props> = (props) => {
   const dispatch = useDispatch<any>()
   const isAccountEnabled = useSelector(selectAccountEnabled())
   const token = useSelector(selectTokenToBridge())
-  const layer = useSelector(selectLayer())
-  const bridgeType = useSelector(selectBridgeType())
   const { amount: recievableAmount } = useAmountToReceive()
   useBridgeSetup()
-
-  // TODO: Move to specific hook add specificity on trigering this hook.
-  useEffect(() => {
-    if (
-      bridgeType === BRIDGE_TYPE.CLASSIC &&
-      token &&
-      token.symbol === 'OMG' &&
-      layer === LAYER.L1
-    ) {
-      dispatch(
-        setBridgeAlert({
-          meta: 'OMG_INFO',
-          type: 'info',
-          text: `The OMG Token was minted in 2017 and it does not conform to the ERC20 token standard.
-      In some cases, three interactions with MetaMask are needed. If you are bridging out of a
-      new wallet, it starts out with a 0 approval, and therefore, only two interactions with
-      MetaMask will be needed.`,
-        })
-      )
-    } else {
-      dispatch(
-        clearBridgeAlert({
-          keys: ['OMG_INFO'],
-        })
-      )
-    }
-  }, [dispatch, bridgeType, token, layer])
+  useBridgeAlerts()
 
   const openTokenPicker = () => {
     dispatch(openModal('tokenPicker'))
